+++ packages/nadle/test/__configs__/task-context-working-dir.ts
@@ -1,10 +1,11 @@
 import { tasks } from "nadle";
 
-tasks
-	.register("withConfiguredWorkingDir", ({ context }) => {
+tasks.register("withConfiguredWorkingDir", {
+	workingDir: "src/test",
+	run: ({ context }) => {
 		console.log(`Hello from ${context.workingDir}`);
-	})
-	.config({ workingDir: "src/test" });
+	}
+});
 
 tasks.register("withoutWorkingDir", ({ context }) => {
 	console.log(`Hello from ${context.workingDir}`);