+++ packages/nadle/test/__fixtures__/cache-dir/with-config/nadle.config.ts
@@ -7,12 +7,14 @@ configure({
 	cacheDir: ".nadle-custom-by-file"
 });
 
-tasks
-	.register("build", async () => {
+tasks.register("build", {
+	outputs: [Outputs.dirs("dist")],
+	inputs: [Inputs.files("input.txt")],
+	run: async () => {
 		const content = await Fs.readFile("input.txt", "utf-8");
 		const modifiedContent = content + " modified";
 
 		await Fs.mkdir("dist", { recursive: true });
 		await Fs.writeFile(Path.join("dist", "output.txt"), modifiedContent);
-	})
-	.config({ outputs: [Outputs.dirs("dist")], inputs: [Inputs.files("input.txt")] });
+	}
+});